Rename to transitionsDisabledReminder
