+++ sentinel/src/schedule/index.js
@@ -42,7 +42,7 @@ const tasks = {
   replications: require('./replications'),
   outbound: require('./outbound'),
   purging: require('./purging'),
-  transitionsErrorReminder: require('./transitions-error-reminder'),
+  transitionsDisabledReminder: require('./transitions-disabled-reminder'),
   backgroundCleanup: require('./background-cleanup')
 };
