+++ sentinel/src/schedule/transitions-disabled-reminder.js
